refactor tests to reduce file size
